refactor: rename mouse-handler module
